refactor: wishlist 로직 ProductSection에서 관리하도록 수정
